+++ src/SeablastPdo.php
@@ -183,11 +183,7 @@ class SeablastPdo extends PDO
     public function queryStrict(string $query, ?int $fetchMode = null, ...$fetchModeArgs)
     {
         $trimmedQuery = trim($query);
-     //   if (!$this->isReadDataTypeQuery($trimmedQuery)) {
-      //      $this->logQuery($trimmedQuery);
-     //   }
-      //  try {
-            // Execute the query based on the presence of fetchMode and fetchModeArgs
+        // Execute the query based on the presence of fetchMode and fetchModeArgs
         if (!empty($fetchModeArgs) && !is_null($fetchMode)) {
             // Spread the fetchModeArgs to pass them as individual arguments
             $stmt = $this->query($trimmedQuery, $fetchMode, ...$fetchModeArgs);
@@ -199,14 +195,13 @@ class SeablastPdo extends PDO
             $stmt = $this->query($trimmedQuery);
         }
 
-        //    $this->addStatement(true, $trimmedQuery);
-
         if ($stmt === false) {
-         ////   $this->addStatement(false, $trimmedQuery); // todo isn't it duplicit error report?
             Debugger::barDump(
-                ['arguments' => func_get_args(),
-                      'errorCode' =>        $this->errorCode(),
-                     'errorInfo' =>    $this->errorInfo()],
+                [
+                    'arguments' => func_get_args(),
+                    'errorCode' => $this->errorCode(),
+                    'errorInfo' => $this->errorInfo()
+                ],
                 'Query failed'
             );
             throw new DbmsException("Query failed.", (int) $this->errorCode());